force use of existing na if pipes are using it
We need to tweak the netmap-adapter selection logic
of netmap_get_hw_na(). If there exists some pipe created
from the adapter, we cannot switch to another adapter.

This is a temporary measure. The ideal solution would be
to not have pipes keep a reference to the parent netmap_adapter,
but only to the proper memory allocator. We can delay
this solution to when we will have proper memory allocator
objects, with their own rerefence counting.