explicit pointers to vp and hostvp
This patch adds to netmap_adapter explicit pointers to the
netmap_vp_adapter's used to connect the na to a VALE switch.
This is necessary to give a uniform access to this information
to external modules.

The nm_bdg_attach() callback now has to initialize these pointers,
instead of returning them via its parameters.