netmap monitors
netmap monitors can be used to do zero-copy monitoring of network traffic
on another adapter, when the latter adapter is working in netmap mode.

Monitors offer to userspace the same interface as any other netmap port,
with as many pairs of netmap rings as the monitored adapter.
However, only the rx rings are actually used. Each monitor rx ring receives
the traffic transiting on both the tx and rx corresponding rings in the
monitored adapter. During registration, the user can choose if she wants
to intercept tx only, rx only, or both tx and rx traffic.

The monitor only sees the frames after they have been consumed in the
monitored adapter:

 - For tx traffic, this is after the slots containing the frames have been
   marked as free. Note that this may happen at a considerably delay after
   frame transmission, since freeing of slots is often done lazily.

 - For rx traffic, this is after the consumer on the monitored adapter
   has released them. In most cases, the consumer is a userspace
   application which may have modified the frame contents.

If the monitor is not able to cope with the stream of frames, excess traffic
will be dropped.

Each ring can be monitored by at most one monitor. This may change in the
future, if we implement monitor chaining.