linux/configure: much faster config
Instead of compiling the tests one at a time, we put all the files in a
directory with a comprehensive Makefile, run 'make -k -j' and then look
for the object files that have actually been created.

Note that the large (x5 and more) speedup is mostly due to the amortized
cost of going through linux kbuild only once, rather than to the -j
parameter.