limit in-kernel use of netmap ring pointers
Now txsync and rxsync callbacks only use the rcur, rhead and nr_hwtail
pointers in the krings, instead of direcly accessing the cur, head
and tail pointers in the rings. The r* pointers are synchronized
with their ring counterparts by netmap_poll() and netmap_ioctl()
as needed.

This improves the reusability of the *xsync callbacks in other contexts,
such as full-passtrough (where we cannot guarantee that the ring
pointers are not accessed concurrently) and the VALE bridge-wrapper
(bwrap).

Before this patch, bwrap had to use the same, shared ring pointers with
both a tx and an rx meaning, confusingly alternating between the two.
With this patch the ring pointers are not touched at all, and the bwrap
simply uses the two separate sets of r* pointers, one in its own krings
and the other in the attached port krings.