linux/make: use the same %.o build rule as Kbuild
Our linux Kbuild file includes the rules for %.o building,
since most netmap files are not in the LINUX directory, and
Kbuild does not support VPATH.

Before this patch, the recipe we were using was not the same
as the corresponding one in linux/scripts/Makefile.build.
This was causing annoying WARNINGs and, possibily, incomplete
rebuilds after file changes.