linux/e1000: removed check on bi->skb
The type of bi has changed from e1000_buffer to e1000_rx_buffer
in 3.18, causing compilation failures.
Since the check on bi->skb was already marked as useless, and this
was the only use of bi, this patches removes the failing code
instead of adding another feature test to configure.