[ptnemtap] Changed handling (enable/disable) kick
In this version we disable guest_[rx|tx]kick in the host, before
to send the kick to the guest. In this way, we avoid to send other
kick before guest disables it.
Guest re-enables kick only in the [rx|tx]sync.
We added usleep() in the kthread, just before to enable host kick,
to wait the guest adds new packet in tx or consumes packets in rx.