make NS_BUF_CHANGED bidirectional
Now the NS_BUF_CHANGED flag is set by netmap whenever
the buffers are changed internally (both pipes and monitors
currently do this).

pkt-gen has been modified so that the "drop copy" optimization
is disabled, on a per-slot base, when the kernel reports a buffer
change.

Note that this dramatically lowers the rate that pkt-gen sees on pipes,
since now it has to touch every packet.