linux/scripts: look for either Makefile or Kbuild
For the purpose of adding a suffix to the patched
linux drivers, we replace the existing driver Makefile
with our own.

The out-of-tree i40e driver does not have a Makefile
inside the i40e subdirectory, but it does have
an equivalent Kbuild file. This patch makes our
scripts to accept either one.