optional exclusive binding of rings
An applications that wants exclusive ownership
of netmap rings may set the new NR_EXCLUSIVE flag
in the flags field of the nmr passed to ioctl(fd, NIOCREGIF, ...).
This combines with the existing options for ring selection, so
that only the requested subset of rings is affected.

Exclusive access fails (with EBUSY) if any of the requested rings
is already bound by a previous ioctl(..., NIOCREGIF, ...).
If successful, future attempts to bind any of the requested rings
(either exclusively or not) will fail with EBUSY until the successful
binding is closed.