change netmap_bdg_learning() to try to avoid hash on MAC src
netmap_bdg_learning() now store into the netmap_vp_adapter(per-port)
the last source MAC address seen to avoid hash on it, because on the
same port is very common to see packet from the same source.

The API is changed because the netmap_bdg_learning() could store it state
into the netmap_vp_adapter.