generic: always keep a reference to the ifp
The generic driver always needs a live ifp to work, but the previous
code held a reference only when the ifp had a native adapter.

Note that netmap applications using the generic driver will now
prevent unloading the nic driver. To avoid this we need to receive
the unload notification from the kernel.