virtio-net: take into account virtio-net header
This patch fixes the virtio-net native adapter in such a way that
virtio-net headers are automatically prepended on TX and RX datapaths.
Two-elements scatter gather lists are used, so that the first element
points to a globally shared zeroed virtio-net header, and the second one
points to a netmap buffer.