generic: keep GRO support enabled on freebsd and linux
This is supported through m_copydata() on freebsd and the corresponding
skb_copy_bits() on linux. These functions have an offset argument that
can be used to copy the skb content over more netmap slots in a very
simple manner. It's not very efficient, since m_copydata() must be
called - with increasing offsets - as many times as the number of
netmap slots needed, and each time it has to skip the data already
copied to get to the required offset. Skipping involves sequential
scanning or lists or array internal to the mbuf/sk_buff implementation.

A more efficient implementation would require a per-os implementation
of the the mbuf scanning, with an external iterator that can be used
to restart the copy from the previous stop (without scanning the same
lists again and again).