lb: added overflow queues
Ask for extra buffers with the -b parameter.
Extra buffers are put in a free queue and used as an extension
of output pipe queues.

The current implementation does not reclaim buffers after they
have been assigned to a pipe, therefore a stalled receiver may
permanently steal all the free buffers.