generic: tx_set_event: fall back to a safe notification scheme
Putting events in the middle between hwtail+1 and hwcur is not
safe in situations where mbufs are destroyed in a
different order than the one they were submitted to the driver.
As an example, this may happen if the internal queue of a driver
gets full (e.g. it happens with a TAP device, where
transmission is implemented by putting the mbuf into an internal
queue to be read by userspace).