generic: fix TX event setup and cleanup
With this patch, with txqdisc == 0, only a single TX event can
exist on each TX ring. We keep track of the mbuf used to implement
the event, since we have to remove the destructor before doing
unregif, in order to prevent crashes due to lazily consumed
mbufs in the drivers.