Keep adaptor's rings either in normal or netmap mode
After this commit, when an adapter is put in Netmap mode, all of its
rings that are not requested to be in Netmap mode are kept in normal
mode, which is they are still attached to the host stack.

Moreover, if a packet is going to hit a TX ring which is in Netmap mode,
the packets will be copied to the RX host ring.