introduce nm_os_{get,put}_module()
In linux these become module_get/put(THIS_MODULE). We need this,
otherwise 'vale-ctl -a vale:eth1', where the driver for eth1 is
not-patched, will leave the netmap module unloadable. Unloading
the module in this state leaks all the memory allocated by netmap.

In FreeBSD these translate to netmap_use_count++/-- as before.