linux: unset the zombie flag on NETDEV_REGISTER
This patch adds a listener to the NETDEV_REGISTER event and unsets the
zombie flag in its handler.

Since commit d8d2569ba01bea1cc5497518125a63370f4551f5, netmap did not
work with LXC containers. That is because to move the device to a new
namespace, it is unregistered first. Unregistering a netmap device
leaves it in a zombie state, which renders it unusable.