add support for batching in nm_dispatch()
The struct nm_pkthdr is extended with more fields including
a NM_MORE_PKTS flag and a pointer to the slot.
The flag is set when there are more packets coming.
The pointer to the slot is also useful to implement zerocopy
(by swapping the index in the original ring)