pkt-gen: ability to generate IPv6 packets
This patch adds two flags -4 and -6 to specify address family. '-4' is by
default. I used the following syntax for IPv6 addresses:

+        *  group@[2001:DB8::1001]:port (w/ brackets and port)
+        *  group@[2001:DB8::1]         (w/ brackets and w/o port)
+        *  group@2001:DB8::1234        (w/o brackets and w/o port)

group is number from 0 to 7, actually it is index of s6_addr16 group in
the in6_addr structure. This group is used for changing addresses.

Also I added checksum adjusting.