transparent forwarding: protect calls to netmap_sw_to_nic()
With this change netmap_sw_to_nic() can be called only when the
user has opened all the hw TX rings. This is needed because to
make sure that TX rings exist and that the user is aware that
is in charge of managing the TX rings (and make sure no other
threads are accessing them concurrently).