linux/ixgbe: use Tx head write back
Reading the TDH was causing Tx hangs when driving a ring
at high speed. With this patch we enable Tx head write back,
which seems to solve the issue.

As a bonus, the CPU utilization is nicely improved.

An implementation very similar to this one was suggested by
Tom Barbette a couple of years ago.