Check if carrier present instead of running.
Replace the check whether interface is running with a check whether
there is actual carrier, otherwise packets will be reported as sent
although there is no actual link on the interface.

Tested by sending packets with:
pkt-gen -i ethX -f tx -n 0 -l 1500

Pkt-gen shows that it is sending packets even after the cable was
physically unplugged from the interface.

The check in ixgbe is the same in txsync handler.