vale: fix crash due to NULL rx rings
Ports attached to a VALE switch may have not have allocated all the
RX rings, therefore the sender must check for the ring
existance before trying to write into it.

Partial allocation may happen if the port has many rings and
only one is opened (-x syntax in nm_open), or if the port
is opened only for TX (/T flag in nm_open).