Add ring range support on NIOCREGIF.
Add new reg NR_REG_RANGE_NIC option which registers a range [start, end]
of rings to the fd on which NIOCREGIF ioctl is issued. _start_ is
provided by the nr_ringid field and _end_ is gotten from the 12 MSBs in
nr_flags. The latter was chosen as it results in the least change to
pass the _end_ of the range.