bwrap: fix ownership management of netmap rings
On attachment to VALE, the usage counter is incremented for the wrapper
rings. Now the increment is propagated also to wrapped port rings.

On detachment from VALE, the wrapper netmap-rings were always
released. Now the ownership is passed to the wrapped adapter, instead.

The previous behaviour did not cause any problem in the normal cases
(attanching/detaching hardware or persistent-VALE ports), but it is
incorrect for netmap-enabled veths, which must manage the netmap-rings
lifetime by themselves.