pkt-gen: let rxseq look at all opened rings
The purpose of this patch is to simplify testing using txseq/rxseq.

Before this patch, rxseq only looked at the first RX ring.  The packets
from txseq, however, may be received on another ring and be completely
missed by rxseq.

Moreover, if the other rings were also opened in netmap mode (as it is
the case with a command like 'pkt-gen -i netmap:eth0 -frxseq'), the
actual destination ring was never emptied. If Flow Control was enabled,
this in turn caused TX-hangs in the sender.