linux/ixgbe: play nice with the ixgbe pointers
The netmap ixgbe driver reuses the next_to_clean pointers
also used by the original driver. With current versions of
the driver, this confuses ixgbe_down().

The previous code reset the pointers to sane values before
leaving netmap mode. This, however, missed the case when ixgbe_down()
is called while some rings are in still in netmap mode.

The current patch always keeps all the pointers updated. This removes
the need for the reset.