mem: use actual number of objects when unmapping
Before this patch, netmap_mem_unmap() was relying on the _objtotal field
for the number of clusters to unmap.  However, the actual number of
objects is stored in the objtotal field, which be less than _objtotal if
allocation failed midway. The lut entries beyond objtotal are garbage,
and therefore netmap_mem_unmap() was sometimes causing panics.