tlem: update producer time before dropping packets
Before this patch, tlem was not updating the producer current time
when trying to remove packets from a full bandwidth-limitation queue.
Packets would then stay in the queue for too much time, causing al lot
of other packets to be unduly dropped. This was breaking the bandwidh
emulation, unless the queue was set to a large enough size to store the
packets arrived between two current-time updates.