linux: virtio_net.c: also drain RX virtqueues when entering netmap mode
The NAPI (used for RX) is disabled when entering netmap mode, so the
driver cannot call virtqueue_get_buf() while the nm_register is detaching
unused RX buffers from the avail ring. However, it is possible that the used
ring has some pending entries (to be consumed by the guest) at that time.
If we don't drain those, detaching the pending avail entries in the RX
virtqueue will result into a "is not a head" bug on the next call to
virtqueue_get_buf().