FreeBSD: Avoid FD leak in stack
In FreeBSD, since compressed mbufs are just consumed and not
visible in soupcall, a zero-length (i.e., event) slot is not
set to the RX ring. As a result, clients are not notified of
incoming FINs thus lose opportunity to close the connection,
leaving it in CLOSE_WAIT state. The only way to solve this
problem is use of per-core lockless buffer that soupcall()
signals to nm_os_st_rx() which runs in the same context.