i40e TX and RX checksum offload
This is Linux specific.
PASTE required
"ethtool -K YOURNIC tx-checksum-{ip-generic (ixgbe), ipv4 (i40e)} on",
but it was only to tell a lie to the stack or it did not actually
offload checksumming. This patch enables actually to offload TX TCP
checksum calculation.  RX side checksum validation in HW is also
enabled by ethtool -K YOURNIC rx on, providing the stack a hint not to
consume time to validate TCP checksum.