update nm_kr_[tr]xempty to check rhead rather than rcur
A TX (RX) ring is empty when head == tail. The cur pointer can be
between head and tail included (typically if the application looks
for more space). In CSB mode, the kernel kloop may see cur ahead
of head because of a (safe) race condition. This was causing
problems with netmap passthrough experiments (over netmap pipes),
where the race condition can be observed.