apps: bridge: use ring->head as ring index
Convert bridge to use ring->head as a ring index, rather than
using ring->cur. This is not strictly necessary in this case, as
for this program ring->cur == ring->head. However, should this
assumption become false in the future, using ring->cur would
be a bug, because ring->cur is the wake up point, and ring->head
is the next packet to process (also see nm_ring_space()).