add per-slot offset support
This patch introduces a new register option: NETMAP_REQ_OPT_OFFSET.
With this option a configurable part of the ptr field in the netmap_slot
can be used to specify an offset into the buffer.
The offset field can be read and updated using the bitmask found
in ring->offset_mask after a successful register.

For RX rings, the user writes the offset o in an empty slot before
passing it to netmap; then, netmap will write the incoming packet at an
offset o' >= o in the buffer. o' may be larger than o because of, e.g.,
alignment constrains.  If o' > o netmap will also update the
corresponding field in the slot. Note that large offsets may cause
the port to split the packet over several slots, setting the NS_MOREFRAG
flag accordingly.

For TX rings, the user may prepare the packet to send at an offset o
into the buffer and write o in the offset field. Netmap will send the
packets starting o bytes in the buffer. Note that the address of the
packet must comply with any alignment constraints that the port may
have, or the result will be undefined. The user may read the alignment
constraint in the new ring->buf_align field.  It is also possibile that
empty slots already come with a non-zero offset o specified in the
offset field. In this case, the user will have to write the packet at an
offset o' >= o.

During open, the user must also declare the maximum offset that she
is going to use. Any offset larger than this will be truncated.