Use size_t instead of u_int for total memory.
In testing, when using buf_num over 420,000 or so, with the default
clusternumbers, we run into integer overflow and segfault.

In our environment, we have 40 core servers (20+20) with the nics using
40 tx/rx rings, with the ring slot size set to 4096. We have to set the
buf_num to about 750000 in order to allocate enough memory for our
application to load with around 100000 extra buffers.

This segfaults without this change.

This is basically adapted from issue #602. I left the max buffers at the
default of 1_000_000 and changed some %lu and %ld to %zu %zd