add kring field to keep track of slots invalidated during NIC reset
Netmap tries to shield applications from NIC resets: during a NIC reset
the hw rings are wiped out and rebuilt from scratch by the NIC driver,
but any active netmap ring is preserved and the new hw rings are
re-initialized using the slots and indices read from the corresponding
netmap ring, if any.

The current implementation, however, unsafely accesses also the part of
the netmap ring in the [head, tail) interval, which belongs to the user.
Proper operation, instead, should only restore the part of the hw ring
that corresponds to the kernel-owned subset of the netmap ring, leaving
the rest for later (i.e., to the *sync callbacks). In preparation for
this, this patch adds a new field to each kring to keep track of the
number of hw rings slots that have yet to be restored from their netmap
ring counterpart.