lock the rings instead of stopping them during NIC reset
Before this patch we stopped the netmap rings when a NIC was going down.
A stopped ring notifies an error to the netmap application the next time
that it will issue a *sync operation. The application is expected to
close the port as a result. This was intended as a proper response to an
'ifconfig down' command, which may leave the NIC down for an unbounded
amount of time.

However, the NIC may also go down and then immediatly up in several
occasions, in particular when the rings of a multi-ring NIC are open in
netmap mode one at time: each new open will cause a down/up cycle which
may cause whomever is using the other already-opened rings to see an
error. In these cases we would rather have the already-opened rings to
never cause an error.

This patch removes the stopped state and just locks the rings when the
NIC goes down. This gives us better behaviour in the down/up scenario,
sacrificing the 'ifconfig down' one, at least for now.