linux/configure: don't pass EXTRA_CFLAGS to external drivers
During the configureation phase, EXTRA_CFLAGS enables -Werror to improve
the sensitivity of the feature tests and selectively disables some of
the warnings to improve the specificity.

The -Werror flag, however, may cause the test build of the unpatched
external drivers to fail, causing them to be unnecessarily disabled.

This patch passes the configuration-phase EXTRA_CFLAGS only to the
feature tests. During this phase the external drivers receive an empty
EXTRA_CFLAGS, thus making the comment in LINUX/default_config.mak.in_
true again.