properly update host-rings mode
Before this patch, the mode (ON/OFF) of all the host rings beyond the
first one for each direction where not being updated during
regif/unregif. This was causing spurious calls to nm_register() which,
for hardware ports, was in turn triggering unnecessary hardware resets.

This affected only applications that were using the "multiple host-rings"
feature.