Fix assymetry in reference counter updates of memory allocator
When a single memory region was provided via the EXTMEM option multiple times, for multiple interfaces, then the memory region was never actually released by netmap. After closing all netmap file descriptors, the external memory remained in use by netmap.

The problem was in netmap_mem_drop: the condition to release the memory reference is when the last user of *netmap_adapter* is gone, not when the last user of *netmap_mem_d* is gone.