pkt-gen.c: update IP/UDP checksums if size changes
When randomizing the packet size using '-l max -l min', the checksums
and IP/UDP lengths are set using the maximum payload size. After that
all size changes are truncations of the payload field. The IP/UDP
lengths and the associated checksums are not recalculated.

The IP header can quickly be modified by changing the length and
modifying the existing checksum (for IPv4). The UDP checksum covers the
entire payload and cannot be trivially updated. The current solution is
to recompute the checksum across the payload each call.

Since recomputing the UDP checksum can have a non-trival cost, I have
added a new option OPT_UPDATE_CSUM (4096). When this option is set,
IP/UDP lengths will be updated and checksums will be valid.