linux: fix to avoid fortify panic on 6.5
Fortify was detecting "buffer overflow in strcpy" in tc_configure()
where a string is copied into a TCA_KIND attribute. The nlreq object has
sufficient space with a 100 character buffer but the attribute pointer
is set to &nlreq.hdr which has a fixed sized. By changing the pointer to
&nlreq, fortify is able to see the extra bytes and not panic.

Also remove a duplicate strcpy() in netmap_sink_init().